Support "enabled" property (defaulting to true).
There are business cases where it might not be desirable to bind a custom domain for every deployment scenario.  For example, one may wish to disable it for certain deployment stages.  This change makes it possible to programmatically disable serverless-domain-manager – for example, in situations where the custom property object is dynamically evaluated based on command line switches or environment variables.