Reverting setupBasePathMapping Hook Change
This appears to have caused a lot of issues for folks, and there are
some good arguments that the package command should not need to pull
down things from AWS to work, and this is really only needed when we
want to deploy the service.

Please see #52 for more information.