Fix getCertArn with wildcard in certificateName
The example in the README.md allows for a wildcard when specifying
a custom domain in the certificateName

```
custom:
  customDomain:
    domainName: serverless.foo.com
    stage: ci
    basePath: api
    certificateName: '*.foo.com'
    createRoute53Record: true
    endpointType: 'regional'
    securityPolicy: tls_1_2
    apiType: rest
```

However, the current implementation throws an error because getCertArn
doesn't strip the wildcard and '.' from the user-specified certificateName.
`certificateArn' is undefined and the `sls create_domain` command fails.

This fix strips '*.' from certificateName if certificateName starts with
'*'.

'*.foo.com' -> 'foo.com'