Fix domain deleting when fail on not exist stack
When stack not exists `findApiId()` fails on `getStack()` and then on `getNestedStack()`. Try block in `removeBasePathMappings()` do not set `externalBasePathExists` flag. As result domain that has other mappings are removed. This fix adds extra check to catch block of `removeBasePathMappings()` if option `preserveExternalPathMappings` is set in stack config (by setting this option user identify that domain should have another mappings) and set the flag `externalBasePathExists` that prevent removing domain.